Test that formatting does not happen when typing on a long line
We currently have a bug caused by the auto formatting that happens when
we reach textwidth. The last two characters are getting swapped before
the line is formatted. This leads to many frustrating situations.

To avoid this situation we want to avoid formatting when we have not
explicitly called a formatting function such as 'gq' in normal mode.