add filetype with lua script for nvim
from neovim 0.8.0, filetype detection is done with filetype.lua instead
of filetype.vim.
The autocommand added in ftdetect/puppet.vim will always come after the
lua detection from nvim and while working fine for simple filetype
assignement, it does not for the subfiletype detection.
As I did not found a way to remove a filetype from lua detection, I had
to replace it with the same logic that was done in the vimscript file.
The autocommand has to be disable to not interfere with the lua detectection.