Rework epp subtype detection
The previous work had some problems:

* it was working for nvim but not at all for vim.
* it did not provide any loop-breaking checks before launching doautocmd
* it was setting `filetype` from within the syntax file instead of the
  ftdetect file.

This reworks the code to fix most of the above issues. The changes are
almost verbatim what Time Pope wrote for the eruby syntax and ftplugin
files in vim/nvim.

With the changes, vim now is able to correctly identify the subtype
based on the file name's extension prior to .epp. This covers most of the
basic cases.

However, I was not able to find a good way to solve finding the filetype
based on parsing file contents and the paths leading to the file. We
don't want to reimplement the whole of builtin file detection, which is
why @shadowwa was trying to reuse the autocommands already determined.

Until we find a way to fix the above issue, we at least now have most of
the filetype detection working, which is better than nothing.

NOTE: I'm intentionally keeping the tests that are failing for vim in
this commit. This way ppl who will investigate this project in the
future will be able to see that the tests actually pass on nvim, but not
on vim.
The next commit will comment out the failing tests though, since
vader.vim still doesn't have a SkipIf instruction.
See: https://github.com/junegunn/vader.vim/pull/201