CI: rename matrix options to more significant names
Currently the test matrix diplays `tests (true`) and `tests (false)`.
This is not super informative and if forces one to dive into the
workflows file to figure out what it means.

With more descriptive names, the purpose of each test becomes more
apparent. Here we're using the value that we need to give to the TESTVIM
env variable when we run the vader test suite.