Prep tests for when SkipIf get merged into vader.vim
I've just made a test run of the PR on vader.vim that adds two new
blocks that makes some tests skippable and it does work! The PR is
incomplete though and it's hard to tell when/if it'll ever get merged.

Nonetheless I'll keep the working test syntax if it can ever come in
hand. The best thing though would be if were able to fix the plugin's
filetype detection for vim, thus removing the need to skip tests.