Avoid overriding formatoptions in puppet's ftplugin
We're using formatexpr. This means that we're explicitly not using vim's
builtin formatting, so the formatoptions are (currently) not used by the
formatexpr function.

If this plugin were to vary the formatting according to formatoptions,
it would actually be up to users to choose what they want to be
formatted, not the plugin itself.