Update keyword list according to puppet 8 documentation
Those new keywords reflect what's currently in
https://www.puppet.com/docs/puppet/8/lang_reserved.html

Most of the keywords added are marked as "reserved" for one reason or
another, so I'm not sure if they'll be used in the future.

`default` was also defined twice, in the keywords and the constants. We
don't need it marked as a keyword since it's a special value.