vader already has a mechanism to mark tests as pending
in the readme I found out that we can add "TODO:" of "FIXME:" in the
description of a Do or Execute block to mark the test as pending. The
readme erroneously mentions only "TODO" and "FIXME" though... it's not
working without the ":".

With this we can include the test result in CI but have it not affect
the exit code. It's not a perfect solution though: we could actually
have those two tests affect the test results for nvim. That conditional
skip though is not implemented. That's what
https://github.com/junegunn/vader.vim/pull/201 exists for.

Still, it's better this way than just not running the tests at all.