syntax: use vimscript syntax directly
The command created in the syntax file is just syntactic sugar to avoid
typing a couple of characters, but it prevents vim from coloring the
lines appropriately.

The amount of characters saved by this shortcut is insignificant.

The command calls also mean if we grep for "def link" we only get to the
command definition and not the actual links.

We'd better have explicit readability rather than saving a tiny handful
of characters.