Start using the syntax group puppetSpaceError
We have a syntax group for highlighting errors of mixtures of spaces and
tabs but it's not being used since it's not linked to any highlight
group.

Now that we're making it visible we can see that it's not working super
great. Let's start testing that it actually works as expected.

The match regexp is only matching one case, where your have a bunch of
spaces and then one tab. The error happens if there's any kind of mixing
between spaces and tabs.

It also needs to be contained inside puppetComment, otherwise it's not
showing up.

Since displaying blobs of red in the editor might annoy some ppl, I've
added a new global option to disable the highlight groups for errors.