syntax: aesthetic refactor: rename syntax groups for bracket/brace/paren
This change has no real effect in practice for highlighting. However, it
makes the names of the syntax groups more logical and more importantly
easier to debug.

Now the regions are named *Block and the start and end delimiters are
named after the symbol that they contain.

Note: the delimiters don't include any spaces anymore. There was no
reason to include that in the delimiters.

The three regions are now made `transparent` since it's not interesting
to highlight what's inside in a special color. That will be done by all
of the other things that they contain.

As a vanity change, I've also flipped the order of the AssertEqual
parameters where I've changed them since they show up in reverse logical
order in vader.vim error messages. I'll adjust all of the tests to use
the same order eventually, but for now I'm changing what I'm touching
anyway.