syntax: vanity change: set container regions as transparent
This does not change how things are shown, but in the same manner as for
the puppetComment match/region in the previous commit, since those
regions' interiors are not set to be highlighted themselves (only their
start and end delimiters are), we might as well set the region as
transparent. The contained elements will be the ones setting visible
elements.