syntax: Add context on all number tests
This looks terrible... but it was super useful when I was removing
things from the number regexps.

vader is absolutely not helpful when a test fails because of an
assertion error. by default it shows the two values that should have
matched, but there's no context of what was tried, or at least the line
number in the file.

The context I've added is very verbose but it helps find exactly what
test failed.