syntax: missing core functions + reorganize groups again
While reviewing all of the new stuff and stuff that's gone, I removed a
bunch of functions instead of moving them to core, like it was done in
the puppet project. We need to bring those back

I've also reorganized again some of the functions and keywords so that
they are in better subgroups. some of the function have an effect on how
the code is traversed so they should be displayed differently. case and
in "choose in function of a variable's contents" so I set them apart.
and and or really are just conditional operators