syntax: use syn-keyword where possibe instead of matching on a single word
Most of the functions and types are currently implemented with syn-match
on a single word. syn-keyword is essential exactly that and we can
combine multiple keywords on one line to make them more compact.

This is essentially a noop but it makes the syntax file easier to read
and maintain.

Some of the function names have `::` in their name, and : is not inside
iskeyword so we can't transform them into keywords. I've kept then words
in alphabetical order per syntax group, so it looks a bit awkward in
some places, but I still think it looks nicer than before.

Note: the matches on the function names with `::` in them currently
don't work: they get matched as puppetName.