Expose formatting bug: comments get messed up alignment by formatting
This bug was reported by russellshackleford in #143

When we provoke vim-puppet to run formatting by typing =>, if we're
placed inside a block of comments that has separate example code
snippets, we would expect the formatting to happen only on the part of
the comment that we're situated in. However that's not what we're
seeing: the hashrockets for the entire comment block are getting aligned
the same level! That can make comments look very malformed, and it's
really annoying to to back from that.

This new test exposes the problem