Skip over commits when aligning hashrockets (Closes: #143)
This solution feels like taking a step backwards. It would be nice to
have some manner of auto-alignment in the code snippets in comments. But
right now the handling of the comment lines is just annoying.

Since we simply skip over the comment lines, we should still be getting
the hashrockets aligned properly "across comment lines", e.g. if there's
one comment line in the middle of a hash, the alignment should work on
all of the non-comment lines in the hash -- provided that the comment
line sits at the same indent level as the rest of the hash after the
auto-indent stage of the formatting function.